pyproject.toml
[project]
name = 'mqtt-automator'
authors = [
    {name = 'Oleg Strizhechenko', email = 'oleg.strizhechenko@gmail.com'},
]
version = 'v0.0.2'
readme = 'README.md'
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = ["mqtt", "aiomqtt", "vakio", "lytko", "yeelink", "self-hosted"]
description = '''Simple smart-home automation system. Self-hosted, LAN only.
Can manage Vakio Base Smart recuperators, Lytko 101 thermostats, and turn on/off Yeelink lamps.'''
dependencies = [
    'PyYAML ~= 6.0.1',
    'aiomqtt~=2.1.0',
    'paho-mqtt~=2.1.0',
]

[project.scripts]
mqtt-automator = "mqtt_automator.automator:main_cli"

[project.optional-dependencies]
test = [
  'pre-commit',
  'pytest',
  'pylint',
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--doctest-modules"
testpaths = [
    "mqtt_automator",
]

[tool.autopep8]
max_line_length = 120

[tool.pylint.'FORMAT']
max-line-length = 120

[tool.pylint.'MESSAGES CONTROL']
# why C0111: because I think module/class docstrings are generally useless, naming should be the documentation
# why use-dict-literal: Empty dict literal may be confusing (set or dict?). Calling to dict is explicit.
disable = 'C0111,use-dict-literal'
